+++ app/jobs/handle_destroyed_booking_job.rb
@@ -1,0 +1,9 @@
+class HandleDestroyedBookingJob < ApplicationJob
+  queue_as :default
+
+  def perform(booking)
+    Bookings::DestroyFirstCheckinMissionJob.perform_now(booking)
+    Bookings::DestroyLastCheckoutMissionJob.perform_now(booking)
+    Bookings::DestroyAllBookedReservationJob.perform_now(booking)
+  end
+end
