+++ app/models/booking.rb
@@ -6,9 +6,7 @@ class Booking < ApplicationRecord
   after_create -> { CreateFirstCheckinMissionJob.perform_now(self) }
   after_create -> { CreateLastCheckoutMissionJob.perform_now(self) }
 
-  after_destroy -> { DestroyFirstCheckinMissionJob.perform_now(self) }
-  after_destroy -> { DestroyLastCheckoutMissionJob.perform_now(self) }
-  after_destroy -> { DestroyAllBookedReservationJob.perform_now(self) }
+  after_destroy -> { HandleDestroyedBookingJob.perform_now(self) }
 
   def self.import(booking_attributes)
     booking_attributes.each do |attributes|
