+++ test/jobs/handle_destroyed_booking_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class HandleDestroyedBookingJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
